fead:Add update and delete function based on User

+++ backend/app/GraphQL/Mutations/FolderMutation.php
@@ -3,14 +3,15 @@
 namespace App\GraphQL\Mutations;
 
 use App\Models\Folder;
-use App\Models\User;
 use Illuminate\Support\Facades\Auth;
+use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
 
 class FolderMutation
 {
-    public function createFolder($root, array $args)
+    public function createFolder($root, array $args, GraphQLContext $context)
     {
-     $user = Auth::guard("sanctum")->user();
+    //  $user = Auth::guard("sanctum")->user();
+    $user = $context->user();
      if(!$user){
         throw new \Exception("ユーザーが認証されていません");
      }
@@ -21,18 +22,31 @@ class FolderMutation
      return $folder;
     }
 
-    public function updateFolder($root, array $args)
+    public function updateFolder($root, array $args, GraphQLContext $context)
     {
+        // $user = Auth::guard("sanctum")->user();
+        $user = $context->user();
+        if(!$user){
+           throw new \Exception("ユーザーが認証されていません");
+        }
+
         $folder = Folder::findOrFail($args['id']);
         $folder->update([
-            'title' => $args['name'] ?? $folder->title,
+            'title' => $args['tile'] ?? $folder->title,
         ]);
         return $folder;
     }
 
-    public function deleteFolder($root, array $args)
+    public function deleteFolder($root, array $args, GraphQLContext $context)
+
     {
-        $folder = Folder::findOrFail($args['id']);
-        return $folder->delete();
+        // $user = Auth::guard("sanctum")->user();
+        $user = $context->user();
+        if(!$user){
+           throw new \Exception("ユーザーが認証されていません");
+        }
+        $folder = Folder::where('id',$args['id'])->where('user_id',$user->id)->firstOrFail();
+        $folder->delete();
+        return $folder;
     }
 }
